requirements.txt
feedparser==6.0.11
google-generativeai==0.8.3
PyYAML==6.0.2
